+++ RELEASE_NOTES.md
@@ -1,3 +1,4 @@
 * add test for Observation conversion from 10 to 40
 * add procedures conversion form dstu2 to r4
 * add medication conversion from dstu2 to r4
+* adding caching for ValidationEngine